--- a/.github/workflows/pytest_postgres.yml
+++ b/.github/workflows/pytest_postgres.yml
@@ -5,51 +5,41 @@ on:
       - master
       - "**dev"
     paths:
-      - splink/**
-      - tests/**
-      - pyproject.toml
+      - "splink/**"
+      - "tests/**"
+      - "pyproject.toml"
 
 jobs:
   test:
-    runs-on: ubuntu-20.04
-    strategy:
-      matrix:
-        python-version: ["3.8.16", "3.9.10"]
+    runs-on: ubuntu-latest
     services:
       pg_splink_ci:
         image: postgres
-        # use default connection options - no need to do any different
         env:
           POSTGRES_USER: splinkognito
           POSTGRES_PASSWORD: splink123!
           POSTGRES_DB: splink_db
-        # Set health checks to wait until postgres has started
         options: >-
           --health-cmd pg_isready
           --health-interval 10s
           --health-timeout 5s
           --health-retries 5
         ports:
-          # map port so we can access
           - 5432:5432
+    strategy:
+      matrix:
+        python-version: ["3.8", "3.9"]
     name: Run tests with Python ${{ matrix.python-version }}
     steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
+      - uses: actions/checkout@v4
+
+      - name: Install poetry using pipx
+        run: pipx install poetry
+
+      - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-
-      - uses: actions/cache@v2
-        with:
-          path: |
-            ~/.cache/pypoetry
-            .venv
-          key: ${{ runner.os }}-py${{ matrix.python-version }}-${{ hashFiles('**/poetry.lock') }}-00
-
-      - uses: snok/install-poetry@v1.2
-        with:
-          version: 1.5.1
-          virtualenvs-in-project: true
+          cache: "poetry"
 
       - name: Install dependencies
         run: |