--- a/.github/workflows/pytest_spark.yml
+++ b/.github/workflows/pytest_spark.yml
@@ -1,4 +1,4 @@
-name: Spark and sqlite tests
+name: Spark tests
 on:
   pull_request:
     branches: