--- a/.github/workflows/pytest_spark_or_sqlite.yml
+++ b/.github/workflows/pytest_spark_or_sqlite.yml
@@ -1,4 +1,4 @@
-name: Run tests with cache
+name: Spark and sqlite tests
 on:
   pull_request:
     branches: