--- a/.github/workflows/pytest_spark_or_sqlite.yml
+++ b/.github/workflows/pytest_spark_or_sqlite.yml
@@ -42,4 +42,4 @@ jobs:
         run: poetry add git+https://github.com/altair-viz/altair.git
 
       - name: Run tests
-        run: poetry run pytest --durations=0 -m "spark_only or sqlite_only" tests/
+        run: poetry run pytest -v --durations=0 -m "spark_only or sqlite_only" tests/