--- a/tests/test_full_example_spark.py
+++ b/tests/test_full_example_spark.py
@@ -16,6 +16,7 @@ from .decorator import mark_with_dialects_including
 from .linker_utils import _test_write_functionality, register_roc_data
 
 
+@mark_with_dialects_including("spark")
 def test_full_example_spark(spark, df_spark, tmp_path, spark_api):
     spark.sql("CREATE DATABASE IF NOT EXISTS `1111`")
     # Annoyingly, this needs an independent linker as csv doesn't