The aim of this PR is:
- to make the overall test suite run faster
- to run the full  duckdb + core tests as fast as possible to speed up finding test failures (to make iterating faster)
 
Results:
- Duckdb and core tests now run in  1m 32s rather than 5m 3s
- Slowest test is now 3m 13s rather than 5m 3s

I have [double checked](https://gist.github.com/RobinL/e3621c50554f4b686ea52769a21e6a87#file-new_duckdb_and_core-txt) the tests which run.  The diff shows that this new structure runs (strictly) _more_ tests than before, because previously we were not running `sqlite` marked tests anywhere.  All previous tests are still run.

Caching of poetry environments is now explicitly supported by the official [setup-python](https://github.com/actions/setup-python/blob/main/docs/advanced-usage.md#caching-packages) action

I have now split out the tests as follows:
- core and duckdb (these all use duckdb where a backend is needed)
- postgres
- spark and sqlite

I have also upgraded to the [latest altair](https://github.com/altair-viz/altair/releases/tag/v5.3.0) (5.3.0) to make sure we use the performance improvements I did to [alt.Chart.from_dict](https://github.com/altair-viz/altair/pull/3383)