@RobinL 2024-03-29
Cache size seems to have ballooned from 400mb to 2gb with these new tests - why?